refactor parser a bit
