+++ packages/plugin/src/types.ts
@@ -2,18 +2,18 @@ import { Rule, AST, Linter } from 'eslint';
 import { GraphQLESTreeNode } from './estree-parser';
 import { ASTNode, GraphQLSchema } from 'graphql';
 import { GraphQLParseOptions } from '@graphql-tools/utils';
-import { GraphQLProjectConfig } from 'graphql-config';
 import { GraphQLESlintRuleListener } from './testkit';
 
-export interface ParserOptions extends Omit<GraphQLParseOptions, 'noLocation'> {
+export interface ParserOptions {
   schema?: string | string[];
-  schemaOptions?: Record<string, any>;
+  siblingOperations?: string | string[];
+  schemaOptions?: Omit<GraphQLParseOptions, 'assumeValidSDL'>;
+  graphQLParserOptions?: Omit<GraphQLParseOptions, 'noLocation'>;
   skipGraphQLConfig?: boolean;
   filePath?: string;
 }
 
 export type ParserServices = {
-  graphqlConfigProject: GraphQLProjectConfig | null;
   hasTypeInfo: boolean;
   schema: GraphQLSchema | null;
 };
